Initial spike of a layered store for binaries.
See http://cdent.tiddlyspace.com/ShuntedBinaryStore and the
README for more details on how it is working.

Some implementation details:

* In this spike we use another text store to store the content. This
  wouldn't be the normal thing, as it is highly inefficient (lots of
  base64), but is good for ease of the spike. For other stores it would
  make sense to write just text to the medium, whatever it might be.

* In tiddler_get when a binary tiddler has its __class__ reassigned
  this is a nefarious hack that only works because we aren't changing
  the structure of the class beyond text and _text.

* Only the policies in the core store matter. The bags created for the
  binary_store are stubs and will not be checked.

* On PUT, tiddler.text is emptied when writing to the core_store.

The test/test_basics.py file just barely proves the concept is workable.
It makes no effort to determine if it is performant or stable in various
environments. Something that will definitely need to be tested is how
it interacts with the Collections system (this is because traversing
collections can require loading from the store). Similarly it will
be important to look at how the tiddlywiki serialization behaves.